Add Mac support in CI, restructure main project
This adds a debug Conan profile for Linux in addition to the existing
release profile, as well as debug and release profiles for Mac. We still
only run Linux in CI though, but it's helpful for local testing. We
assume x86_64 on Linux and armv8 on Mac.

We also add OS detection to the build-and-test.sh CI script.

Finally, the main executable is extracted to a separate project, and we
run it in CI just to see that it got built and can run. No further
checking is done.