Make main program read a file
Just lex it and print the tokens, since that's the only thing we can
print atm. Run this in CI as a sanity check.

I was going to use clipp for cmdline parsing, but it's broken as of
C++20: https://github.com/muellan/clipp/issues/53. It also seems to be
unmaintained, the PR that fixes it has been open for >2 years:
https://github.com/muellan/clipp/pull/54

Will need to look into finding a new cmdline parsing library.